+++ ProjectBane/Bane-Platform/Platform-Windows/Input/InputMouseDevice.cpp
@@ -7,10 +7,10 @@ struct InputMouseDevice::Impl
 {
 	Impl(Window* pWin)
 	{
-		WindowSize = float2(static_cast<float>(pWin->GetWidth()), static_cast<float>(pWin->GetHeight()));
+		WindowSize = double2(static_cast<double>(pWin->GetWidth()), static_cast<double>(pWin->GetHeight()));
 		POINT Cursor;
 		GetCursorPos(&Cursor);
-		MousePosition = float2(static_cast<float>(Cursor.x), static_cast<float>(Cursor.y));
+		MousePosition = double2(static_cast<double>(Cursor.x), static_cast<double>(Cursor.y));
 
 		BANE_CHECK(
 			SUCCEEDED(
@@ -39,10 +39,10 @@ struct InputMouseDevice::Impl
 	}
 	IDirectInput8* DirectInput;
 	IDirectInputDevice8* Mouse;
-	float2 MouseDelta;
-	float2 MousePosition;
-	float2 LastMousePosition;
-	float2 WindowSize;
+	double2 MouseDelta;
+	double2 MousePosition;
+	double2 LastMousePosition;
+	double2 WindowSize;
 	float MouseScroll;
 	bool bButtons[5];
 	bool bLastButtons[5];
@@ -54,7 +54,7 @@ struct InputMouseDevice::Impl
 		Mouse->Acquire();
 		Mouse->GetDeviceState(sizeof(DIMOUSESTATE2), &MouseState);
 		LastMousePosition = MousePosition;
-		MouseDelta = float2(static_cast<float>(MouseState.lX), static_cast<float>(MouseState.lY)) * (WindowSize.y / WindowSize.x);
+		MouseDelta = double2(static_cast<double>(MouseState.lX), static_cast<double>(MouseState.lY)) * (WindowSize.y / WindowSize.x);
 		MousePosition += MouseDelta;
 		MouseScroll = static_cast<float>(MouseState.lZ);
 	}
@@ -71,12 +71,12 @@ InputMouseDevice::~InputMouseDevice()
 	delete m_Impl;
 }
 
-float2 InputMouseDevice::GetMouseDelta() const
+double2 InputMouseDevice::GetMouseDelta() const
 {
 	return m_Impl->MouseDelta;
 }
 
-float2 InputMouseDevice::GetMousePosition() const
+double2 InputMouseDevice::GetMousePosition() const
 {
 	return m_Impl->MousePosition;
 }
